Split up ptvsd._main.
